+++ app/javascript/components/workflows/proctor/exams/index.tsx
@@ -63,8 +63,7 @@ import { sendMessage } from '@hourglass/common/api/proctor/messages/create';
 import './index.scss';
 import { BsListCheck } from 'react-icons/bs';
 import { doFinalize } from '@hourglass/common/api/proctor/exams/finalize';
-import './newabove.scss';
-import Tooltip from '@hourglass/workflows/student/exams/show/components/Tooltip';
+import { NewMessages, PreviousMessages } from '@hourglass/common/messages';
 
 export interface MessageProps {
   icon: IconType;
@@ -587,48 +586,6 @@ const SingleMessage: React.FC<{
       throw new ExhaustiveSwitchError(message);
   }
 };
-
-export const NewMessages: React.FC<{
-  onClick: () => void;
-}> = ({ onClick }) => (
-  <Tooltip
-    message="Click to dismiss"
-  >
-    <Button
-      block
-      variant="outline-light"
-      className="text-dark"
-      as="div"
-      onClick={onClick}
-    >
-      <Row className="m-0 fancy-hr">
-        <Col>
-          <hr className="fancy-left-warning" />
-        </Col>
-        <Col sm="auto" className="p-0 d-flex justify-content-center flex-column">
-          New Messages
-        </Col>
-        <Col>
-          <hr className="fancy-right-warning" />
-        </Col>
-      </Row>
-    </Button>
-  </Tooltip>
-);
-
-export const PreviousMessages: React.FC = () => (
-  <Row className="m-0 fancy-hr">
-    <Col>
-      <hr className="fancy-left" />
-    </Col>
-    <Col sm="auto" className="p-0 d-flex justify-content-center flex-column">
-      Previous messages
-    </Col>
-    <Col>
-      <hr className="fancy-right" />
-    </Col>
-  </Row>
-);
 
 type FilterVals = { value: string; label: string; };
